Refactor authentication logic by consolidating authOptions and removing unused code
